Feature - Single Port Mode (#1)
* Single Port Mode

Single port mode multiplexes all client communication onto the server's main listening port. Tracks connections with a map of strings to channels. The string representation of the remote address is
used as the key. ACK, DATA, and ERRORs are sent on the connection's channel. Calls to conn.readFromNet read off the channel instead of a network connection.

* Clean-up duplicate code. Add timeouts and error handling.
* Read data directly into conn.rx instead of through conn.buf
* Prevent creating a new timer on everyread from reqChan
* Refactor incoming datagram dispatch
* Improved handling out of sequence ACKs.
* Close connection before removing addr from reqMap.
* Correct existing tests
* Test client against single port server
* Prevent Travis CI using uppercase vCabbage in GOPATH
* Update README with information about single port mode
* Correct misspellings